Dialyzer fix for riak_kv_map_localphase.erl:
riak_kv_map_localphase.erl:56: Call to missing or unexported function
riak_kv_vnode:map/5

I think this is a holdover from the edit that did away with
riak_kv_vnode:map/5 entirely.  It used to be defined like this:

  map(Preflist, ClientPid, QTerm, BKey, KeyData) ->
      riak_core_vnode_master:command(Preflist,
                                     ?KV_MAP_REQ{
                                        qterm=QTerm,
                                        bkey=BKey,
                                        keydata=KeyData},
                                     {fsm, undefined, ClientPid},
                                     riak_kv_vnode_master).

... and it was fed by:

  src/riak_kv_phase_proto.erl:    gen_fsm:send_event(PhasePid, {input, Inputs}).

... but that code is now missing from riak_kv_phase_proto.erl.
For additional dead-code argument, this would cause a very noticable
runtime crash if it ever happened during testing or in production.