Change capability() callback arity from 2 -> 3.
It hit me that the capabilities of an instantiation of a
riak_kv_multi_backend driver depend on the configuration of the driver,
which is based on bucket names.  Therefore, to ask an instantiation for its
actual capabilities, we need to include the bucket name.

In theory, the riak_kv_multi_backend driver will disappear soon.  But I
don't know when, so we'll assume the worst: that it will live forever.