I think this finishes the ETS backend, hooray (and !@#$! about time).
Added fold_test() EUnit test to explicitly check for correctness of
general fold behavior.  (There was no such test before now.)  The ETS
backend is passing that new test.  Note that there's a reliance on a
change (same branch name as this one) over in riak_core that adds a
new vnode request, #riak_core_fold_req_v2.

I believe but haven't 100% verified that all v1 core fold requests
should just work as-is.

Also, there's a bugfix in here for iterators that are operating on
an ordered bucket and also have a specific bucket to operate on.