reduce the size of {Bucket,Key} serialization.
instead of term_to_binary, take advantage of knowing that
we always have a 2-tuple of binaries, and encode accordingly.

this brings the overhead per item in bitcask from 13 bytes to 3.

detect via an initial flag byte whether a key was stored via
term_to_binary or the new method, for partial backward compatibility.

not entirely sufficient on its own for a "mixed" bitcask with some
keys stored using the term_to_binary encoding, as the lookup hash key
in bitcask's keydir will differ for the writes before and after this.