Add range op to leveldb backend
Add ability to retrieve a range of `{Key, Value}` tuples based on
`Start` and `End` keys, inclusive.  The current API also requires a `Limit`
to avoid streaming back too much data.

This is the foundation of a native range operation for Riak when using
ordered data.  I'd like to see the range written as a NIF in the eleveldb
driver but for now this is just fine.