Replace call to crypto:rand_uniform with random:uniform.
We have observed a crash in crypto:rand_uniform on Solaris.  Post 0.14.2
we have changed the build process for building the embedded erlang
packaged in the riak release and how openssl is linked.  Additionally,
the code for listing keys has been reworked on the 1.0 branch and no
longer requires a seed for the ebloom call.

Lowest risk option is to replace the calls to crypto:rand_uniform
with random:uniform instead.

Fixes: zd://865