Add validation for query params
Note: I didn't bring in Erlando just to be Joe Cool.  Writing functions like
`malformed_request` is always a PITA because you end up nesting case
statements.  In the past I've fought this by using foldl and supporting
funs but it still feels clumsy and it's basically implementing what a monad
gives me for free -- the ability to control execution.  At the first instance
of a malformed param I just want to give up and return an error and I want
to do this without writing ugly nested case statements.